DiagnosticsView: add actions to the Tools menu
Grouped in a KMenuAction "Diagnostics View". Otherwise those actions
aren't discoverable unless you happen to see them in the
configure-keyboard-shortcuts dialog.

Add a ui.rc file for diagnosticsview.

Add icons for a the rest of the actions.

Associate the actions' shortcuts with the diagnostics view and the
viewmanager, but not the mainwindow (like what is already done for
other components/plugins).