gitblame: Fix inline notes with rtl text
With rtl text, the inline notes appeared over the text making it very
hard to work with rtl text in such circumstances. To fix this, api
in KTE has been modified to provide layoutDirection information of the
current line. A provider has to take into account this information when
painting the note.