Add LSP-settings for XML
Install instructions

The easiest way to get executable for this language server
is to download the prebuild binary from the link in the
URL included. Downloaded .zip file needs to be unpacked.
Binary name contains OS type as suffix like "lemminx-linux"
and needs to be renamed to just "lemminx" and placed
somewhere in $PATH.