DocManager: Also listen to documentUrlChanged signal
Currently we only listen to ReadOnlyPart::urlChanged signal. This can
lead to issues where we might end up with multiple documents with the
same url open in a single viewspace.

One of the case I found is:
- open Kate with a project e.g., `kate -b -n /path/to/repo`
- open a file in the project
- use the urlbar to open another file in a _different_ project

=> You get the same file open twice.

The reason is that the part signal fires later and someone (view
manager) requests a document for the new url before that and we end up
with 2 views for the same document.