Allow reordering sidebar tabs using DND
Much easier than what we currently have which is to shuffle stuff until
it matches what you want.

Reordering is limited to current sidebar and current MultiTabBar. This
simplifies the implementation. This means that you can only reorder tabs
within a sidebar. You can't drop a tab in a different sidebar at an
exact location. In most cases, the user will not be shuffling around
things constantly so I think this will be enough.