dap/client.cpp: Use QJsonObject instead of null
GDB has now `dap` interpreter, however it did not work with Kate due to the request being wrong.

I asked in their mailing list if it's a bug in GDB DAP or Kate here: https://sourceware.org/pipermail/gdb/2024-January/051058.html

Instead of sending null request, send object in the request. This makes GDB with `dap` interpreter work. It is also how the request should be made in the first place: Some debug adapters do not care if it's null or object, but may as well return the correct value. :smile: