a11y: Set buddy for icon size label
Specify the spinbox as buddy of the
"Icon size for left and right sidebar buttons" label
in Kate's
"Settings" -> "Configure Kate" -> "Behaviour"
config page.

This allows using the accelerator to jump to the
spinbox and also makes sure the label-for/labelled-by
relationship get set on the accessibility layer,
so the Orca screen reader announces the label as well
when the spinbox gets focus, so it's clear what this
is actually for.

While at it, also add a colon at the end of the label,
to align this with how it's done for the other labels.