a11y: Set buddy for "Diff Style" label
Specify the combobox as buddy of the
"Diff Size" label in Kate's
"Settings" -> "Configure Kate" -> "Behaviour"
config page.

This allows using the accelerator to jump to the
combobox and makes sure the label-for/labelled-by
relationship get set on the accessibility layer,
so the Orca screen reader announces the label as well
when the combobox gets focus, so it's clear what this
is actually for.