a11y: Improve tab order for "Session" config page
Align the tab order (i.e. the order in which
widgets receive focus when moving through
the dialog using the Tab key) with the visual order
for Kate's "Settings" -> "Configure Kate" -> "Session"
tab page so the focus doesn't unexpectedly jump around.

To achieve that, just remove the explicitly
set tab order that didn't include all of the
widgets. This makes the tab order the same as the
construction order, which aligns with the
top-to-bottom order of the widgets as expected.