Refer to qml language server binary by upstream name
The upstream name in Qt is just `qmlls`. Distros may patch it to be
`qmlls6` or `qmlls-qt6`, but we should try to ask for the official name
here. Ideally any distro pacthing would take the form of symlinks so the
original name would still work. If this isn't the case, it should be
probably be fixed by them.


(cherry picked from commit eb4380073f6a1ebddfc0887e7ec0b7b68979f63a)