cmake file API: improve handling of compile_commands.json
Under Linux, crete a symlink of the compile_commands.json file. If
you have only one build tree, this will work and always be current.
If there was already such a symlink, pointing to somewhere else, than that somewhere else
will be current. If it already exists as a file, nothing will be updated.

Under Windows, always copy the current compile_commands.json.