Add GUI tests to Kwrite
**Adds GUI tests to Kwrite.**

This MR is based on the work done in Dolphin appium tests. The main difference in Kate is that there are tests run before the install. To run appium tests, we need the app to be already installed.
I had to change kde-ci-yml to run the tests after install.

```
Options:  
  test-before-installing: False  
  run-tests: True  
```

I do not know what this change implies for Kate and if this can be accepted.
Other possible answers : 
- update Appium to allow it to run tests without the installed app (not ready and a lot of work according to Harald)
- update the CI to run some tests before and some tests after the install