plugin_kategdp: When breakpoint mark changes, toggle breakpoint for it
Breakpoint marks were currently just visual, they did not actually do
anything when it came to toggling the breakpoints.

This makes sure whenever the document updates a breakpoint marker,
it's actually toggled. If debugger is running and you set markers to file `foo.cpp` then close the file,
and reopen it, Kate asks the debugger if the file has any breakpoints and if so, it will mark them back.

There is also a change for the toggle breakpoint keyboard shortcut, where
it toggles the breakpoint using the marker system, instead of having two 
competing systems.

BUG:481691