Use documentation from completionItem/resolve
Some language servers (e.g. haskell-language-server) do not send full data for a completion immediately, and send only a stub for e.g. documentation.

This MR intends to get those completions into Kate. That said, it still currently has quite a lot of issues:
- It resets the completion widget completely each time, since dataChanged did not cause the view to update.
  - This probably requires some KTextEditor-side changes, that I have no idea how to do.
- HLS returns Markdown, but Kate renders it as plain text currently
- Looking at VSCode, it seems like the resolved docs are supposed to be appended to the original docs?
  - Maybe I should add a `bool resolved` member to completion items?
- I got a crash at some point, probably a missing cancel for the request.

Hence this MR is a draft, but I did decide to open it for visibility and maybe some help :smile: