build: Add flutter support
Atm it only has the following features:
- targets detection/creation by checking if the current project is flutter
  or if the project has a .vscode/launch.json file containing flutter targets
- hot reload
- hot restart

Hot reload/restart is done via buttons that appear once a flutter app is
running. However, they don't hide once shown as it is very difficult to
tell if the flutter app has stopped. One can continously poll the terminals
on linux to do this, but it still won't work as on windows the Terminal
has no information on foreground process.

More features are planned, but I am still not 100% sure about this
approach so consider this an rfc atm.