Load MDI ToolView shortcuts as custom to avoid unintentional reset
User-customized shortcuts in the toolviewmanager section, written to katerc, were then being loaded on the next instance of opening Kate as default shortcuts. This was causing KActionCollection::writeSettings to remove them if the user opened Customize Keyboard Shortcuts again and clicked OK, as it saw it as a default shortcut that was unnecessarily stored in the config file. This proposed merge changes from loading the stored settings as defaults in the ActionCollection, to loading them as just plain shortcuts directly to the individual KToggleActions/QActions (before they are then added to the ActionCollection). It also makes that conditional on the shortcut existing/not being empty, as otherwise this function would fill the config file with shortcut="none" lines for every unassigned toolviewmanager action.