LoadInKDevelop: change the type from boolean to string
A string LoadInKDevelop JSON value is more flexible and allows to
prevent loading plugins in KDevelop versions that lack necessary support
for them. Judging by the implementation of
KPluginMetaData::value(const QString &, bool), KDevelop versions that
expect a boolean LoadInKDevelop value would load a plugin with a string
value if and only if it equals "true".

The implementation of
KPluginMetaData::value(const QString &, const QString &) converts a
boolean value to string but also prints a warning about the unexpected
type. So while the LoadInKDevelop values of plugins compatible with any
KDevelop version could remain boolean, it is better to convert them all
to string in order to prevent warnings in newer KDevelop versions.