QuickOpen: List open projects and allow switching to them from quick switcher
This lists the available projects in the quickswitcher, then if user
selects one and opens it, we tell the project plugin to switch to that
project.

This should make it easier to switch between different projects, especially if you have a lot of them in your session.

The behavior is exact same as selecting the project from the combobox list.

| Before Selection | After Selection |
| ------ | ------ |
| ![image](/uploads/1b4b12c063d3da6d9c18bed448bb96ed/image.png){width=1630 height=1067}       | ![image](/uploads/97b456f267a9166b65e8e929c8802d31/image.png){width=1630 height=1067}       |