dap: Dont require 2 clicks to stop the debugger
When I click stop, I want it to stop and not get stuck in some
intermediate state.

The outputText is not cleared on stopping now so that you can see the
output even after the debug session ended. There is no point in clearing
it on stopping imo. Once a new debug session starts the output is auto
cleared then.

This also makes the PostMortem state useless so that is removed