Show git status with file name in project tree
Files in project tree view now show their status in git. The existing
git status data is reused for this. I had to make the git widget a bit
less lazy to make this work correctly.

The status is shown minimally i.e., a small circle is drawn in front of
the file name. If the file is modified the circle is red, if the file is
staged its green. The parent folders of the said file show the circle in
front of their name as well. I haven't included Untracked files yet as it might
slow things down, but we'll see. Perhaps it makes sense to add Untracked
files too.

BUG: 489696