gitwidget: Update git widget with correct folder on project change
Fixes a regression where changing the project does not update the
gitWidget with correct folder, showing only files of the first project
opened. This worked previously but seems to have broken when the git widget combobox was removed.

This adds a new method that can is called from kateprojectpluginview,
which both sets the dotGitFolder and then refreshes the view.