Fix snippets not working unless completion is hard invoked
They should not be completely disabled unless someone hard presses
ctrl-space as it makes the experience very annoying. Overtime one can
get used to the annoyance but I am not sure its a good idea.

The new logic honors ktexteditor minimal word length and starts the
completion if the last char is a letter. This should avoid the issue
with snippet completion starting on pressing `.`. However, this might
still be too simplistic but we can refine on it instead of hard
disabling the feature

BUG: 495416
CCBUG: 462383