Skip "Save macro recording" action when recording is going on
It is not supported, the user must end the recording by either playing
the macro or explicitly ending the recording before trying to save it.

With this commit, we warn the user about this and avoid recording the
shortcut.

BUG: 494464
(cherry picked from commit 11a1e88ff158af88f082f3bccb7ee281712db5f0)