Revert "Use path_helper utility on macos for setting PATH"
Using path_helper doesn't really fix the issue completely. The user will
still need to provide all the paths in /etc/paths.d that they want made
available in Kate.

Also, path_helper seems to be a shell util and the manpage says that
only shells should rely on it.

To fix this issue, we must provide a way to modify the PATH ourselves

This reverts commit 51097a672e627d776298b06e558104fb1224682e.