Allow the user to add paths to PATH var
Kate relies heavily on external executables for a lot of functionality
for e.g., LSP servers, formatters, linters, git etc. However, we dont
provide any way to the user to provide a path where these programs can
be found. As such the user has only 1 way to fix this which is to modify
some global file and ensure the PATH is set there.

On linux, things are messy in general so it kind of works sometimes. On
windows you can change your PATH variable in settings. On mac, you can
sort of edit /etc/paths. The problem with these approaches is that it
forces the user to change things globally even if this functionality is
only required by Kate on the whole system.

This patch introduces a way to prepend paths to the existing PATH env
var.

BUG: 490926