Control stack frame and scope from UI
Currently the backend tries to control the UI and UI tries to control
the backend and it leads to a somewhat annoying experience. For e.g.,
- I switch stackframe while I was looking at "Locals" scope
- After switching stackframe, the scope defaults to the first value
(Arguments in GDB usually). So I have to change the scope to Locals
every single time.

To improve things, we do the following:
- The UI tells the backend about current stack and scope
- The scope is retained. The logic for this currently is extremely
simple, we just set the QComboBox::currentIndex value as the current
scope.
- UI is now updated less often resulting in a slightly smoother
experience. For e.g., after changing the stackframe, the new stackframe
will still be selected.