dap: Improve terminate request handling
Currently we dont handle terminate request response at all and assume
a "terminate" event will always come but this doesn't always happen. For
example, if you specify a non-existent program path to gdb it will fail
to start the program but gdb itself will continue to run without
reporting any error. Trying to terminate/stop the debugging session
after that doesn't work because the gdb-dap never sends a terminate
event (as there is nothing to terminate).

This commit handles terminate response and emits debugeeTerminated
signal. The DAP spec says that if termination fails, a client should
try to force shutdown by sending a disconnect. We do this as well now.

Aside from that, the "Terminate" state is removed as its no longer
needed anywhere else.