Implement LSPClientConfigPage::defaults() correctly
Functions that override KTextEditor::ConfigPage::defaults() are empty or
simply call reset() in many KTextEditor and Kate classes. This is not a
problem for Kate, KWrite, RKWard, Kile or KTechLab, because these
applications never invoke defaults(). KDevelop, however, invokes
defaults() when the user clicks the Restore Defaults button in the
Configure KDevelop dialog.

Instead of calling reset() in LSPClientConfigPage::defaults(), set the
checked state of each checkbox on the Client Settings tab of the LSP
Client configuration page to the default value of the corresponding
option as defined in LSPClientPlugin::readConfig().

Do not clear the empty-by-default UI elements on the second and third
tabs of the LSP Client configuration page to avoid accidental user data
loss. KDevelop's Clazy and Clang-Tidy configuration pages similarly do
not remove check sets configured by the user when the Restore Defaults
button is clicked.

Extract struct LSPClientPluginOptions and initialize its data members to
the actual default values defined in LSPClientPlugin::readConfig() so as
to reuse the default values of the LSP Client options and thus prevent
them from getting out of sync. The extracted struct also allows to share
between LSPClientConfigPage::reset() and LSPClientConfigPage::defaults()
the code that populates the checkboxes.

Reorder the declarations of the data members moved into
LSPClientPluginOptions and the assignments to them in
LSPClientPlugin::readConfig() to match the order in
LSPClientConfigPage::apply() and LSPClientConfigPage::reset().

Make LSPClientPlugin::readConfig() private and document the assumption
relied upon by the updated definition of this function.