Pass `this` as context in SnippetRepository()
8f2f2428c02fc2d44e374672512e9beb60cfab81 removed QObject from the list
of base classes of SnippetRepository and started passing model() instead
of `this` as the context to an asynchronous invocation of a lambda in
SnippetRepository(). QStandardItem::model() returns nullptr in
SnippetRepository() because the item is constructed first, then added to
a model. So the context of the invocation effectively became null in the
aforementioned commit.

As a result, if RKWard is installed in the system, KDevelop's
test_ktexteditorpluginintegration segfaults with the following call
stack:
<one of the accesses to SnippetRepository::m_file>
SnippetRepository::parseFile()
<Qt event-loop-related functions>
QTest::qWait(1) [the first call, before self()->shutdown()]
TestCore::shutdown()
TestKTextEditorPluginIntegration::cleanupTestCase()

The test crashes only if RKWard is installed because no other
application installs its snippets into
/usr/share/ktexteditor_snippets/data. User-defined snippets in the home
directory are not found by the calls to QStandardPaths::locateAll() in
SnippetStore() because of the call to
QStandardPaths::setTestModeEnabled(true) in
KDevelop::AutoTestShell::init(). And neither SnippetRepository() nor its
asynchronous crashing lambda are invoked when no snippets are found.

Restore QObject as a base class of SnippetRepository and pass `this` as
the context again. Now the invocation of SnippetRepository::parseFile()
is canceled when the SnippetStore (model) instance is destroyed along
with all its SnippetRepository items during the destruction of the
KTextEditor plugin in TestKTextEditorPluginIntegration::testPlugin() and
TestKTextEditorPluginIntegration::testPluginUnload().

Replace QTimer::singleShot(0, ...) with
QMetaObject::invokeMethod(..., Qt::QueuedConnection). The former invokes
the latter after some string manipulation, so this is an optimization.