diagnostics: store QUrl in UserRole and retrieve consistently
... since retrieving a QUrl from a QString leads to an empty QUrl,
which breaks file-specific local suppression, as the empty url does not
match the specific file.

Using QUrl is preferred over QString since the former is used throughout
the KDE API as well as in LSP protocol, and it also preserves more
information than might only be in a local file path.