Fix cursor position not being restored with multiple views
The session autosave that we trigger when starting up should specify
autosave=true, otherwise during session save we delete the session
config groups when saving data.

Add a test for this case, it starts up the app then opens a session
with 2 docs and then sets cursor position. We then close the app and
start up again similar to how the app is started in real world. Then
we check whether cursor position is correct

BUG: 506956