Add ability to unflatten the path, add files
This adds the ability to add file/folder to a part of flattened path.

e.g for a flattened item = parent/child1/child2

A user might want to add a file 2 "parent".

This is done by "unflattening" the tree, which is essentially rebuilding
the tree, adding the file, and flattening again.

To make it a bit responsive, the path portions are underlined when
you hover over them.