Add a test for KateViewSpace tabbar context menu
It was missing and will help ensure we don't break things. Some things
are already broken for e.g., the context menu for "widget tabs" doesn't
work.

To expose the context menu a new function buildContextMenu is introduced
which allows testing the context menu easily. This function restructures
things a little bit and does things differently for e.g., every action
now connects to a slot instead of exec'ing the menu and then matching an
action.