Fix passing arguments to DiffParams
The arguments must be the same as what was passed when setting up
the git process. This is because in Flatpak world, the process gets
executed by another process with its own args prepended to our args.
When we later try to use the args from DiffParams, things break because
the args are no longer git args, they are flatpak-spawn args + git
args.

To fix this, always store the args that we used the first time.