Fix infinite loop in git blame
It was assuming that the commit hash is always arg[1] which is not
true in the case of flatpak where the process is invoked via a helper.

So instead, use better logic to find the commit that the process
result is for.

BUG: 508577


(cherry picked from commit 537542367a509055c73670da94feceac6b4add59)

Co-authored-by: Waqar Ahmed <waqar.17a@gmail.com>