Ignore unix signals when quitting
If a signal comes after we have quit and closed the mainWindow, we try
to save the session again. When saving a session we clear old saved
session data and then save the session. But if this is done too late,
then we end up saving nothing and loosing all data in the session.

BUG: 508494