Improve pre context string limit logic
Dont limit the precontext to 25% of available width always. It results
in a lot of waste and annoyance because in many cases there is a ton of
space available but the search delegate just ignores that. Pre context
is often important in understanding the matched line better.

The new logic reduces the pre context string if both of the following
are true:
 - the total text width exceeds the available width
 - pre context + match width exceeds the available width