Simplify local variables code
No need to emit a signal per variable, we can emit a single signal and
handle all the data in LocalsView class. Also remove the open/close
variable scope signals, they are useless now.

This change makes it possible to combine multiple scopes if needed for
e.g., Locals & Arguments. It also makes it possible to do insertion in
a more optimal way i.e., inserting a bunch of items into the tree
instead of one at a time.