dap: be aware of some incompatible servers
Currently multi-session debugging is not supported, so the reverse
request `startDebugging` is rejected.

Server like [vscode-js-debug](https://github.com/microsoft/vscode-js-debug)
relies on `startDebugging`, but it ignores the client's capabilities
and the session is stuck and can not be terminated.

Since they refuse to [follow the specification](https://github.com/microsoft/vscode-js-debug/issues/2156),
this patch should detect the reverse request and terminate
the unsuccessful session.