bookmarksplugin: fix warning about widget already having a layout
'QLayout: Attempting to add QLayout "" to QWidget "", which already has
a layout'

Passing a QWidget* to the constructor of Q{V,H}BoxLayout sets that
layout as the layout of the that widget (a widget can only have one
layout). The layout, hLayout, is reparented by the call to
insertLayout().