Improve LocalsView UX
Whenever the user steps into/over/out, we re-request scopes and
variables again. We then clear the treeview and rebuild it with new
data. This works, but has terrible UX because the tree collapses every
time the user steps through code make the whole feature close to
unusable.

With this change, instead of clearing the whole tree we replace
existing items and remove extra items. This makes the view stable, no
automatic collapsing of items.

Note that the reason we re-request scopes again is because of the DAP
protocol. The protocol mandates that whenever a "stopped" event is sent,
a client must re-request threads, stackframe and scopes again. Thus we
need to handle all of this on the client side.

Also note that some existing signals and apis from backend are removed.
E.g changeScope, which was mostly useless because there is no such thing.
We are always requesting variables for variableReference. Similarly
LocalsView visibility was unused so it has been removed.